+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.19.0-rc.1] - Unreleased
+## [0.19.0-rc.1] - 2025-09-24
 
 - Support Bevy 0.17.0-rc.1
 - Rename `EnhancedInputSet` to `EnhancedInputSystems`